Marks lowered parameters as such
Parameters to a __local memory area are managed by the VC4CL runtime
by creating a temporary buffer and passing the address to that.
If we can lower the memory into VPM, we also don't need to reserve
a temporary buffer for that. But we still need all the parameter information
in the VC4CL runtime. So this simply adds a flag telling the runtime
that this parameter does not need a temporary buffer.