Moves SIMDVector out of Value
Value no only holds a pointer to a globally stored SIMDVector.
Similar to complex data types, SIMDVectors are stored in the module as well
as globally, are immutable and duplicate entries are merged.

For the emulator, which highly utilizes SIMDVectors, the vectors are handled as
value-types, identical to how they were handled before.

Effects:
- reduces size of a single Value for 32/64-bit systems from 136/144 to 16/24 Byte!
- no noticable changes in compilation performance (execution speed)