Converts LongLocal to additional LocalData
The LocalData type is a base for more advanced data for single Locals.
Since it is intended to be set for only a small subset of all locals
(and required to be polymorphic) it is implemented via an indirection
and not as part of the local itself.

This change also allows Parameters to be "marked" as 64-bit local
and have upper and lower part (required for non-kernel functions to
be inlined).

The LocalData type will also be extended to store some other more
advanced Local data including the reference. Thus in the long run
we save up to 1 word of memory per local.