Adds experimental optimization step to cache memory in VPM
The optimization will pre-load (currently not implemented) and write-back
the contents of the VPM cache area from/to RAM at the beginning/end of the
work-group by a single work-item while all "access" is done on the VPM cached
data.

Although this optimization passes for most of the emulation test-cases,
it is still very unflexible in regard to memory range accessed and the
actual performance gain so far seems to be minimal.