Fixes logging for parallel compilation
Fixes a race condition when multiple entry-points of the VC4C API are
called in parallel (e.g. parallel compilation) and allows the logging
to be forwared to the correct output.

Also fixes the tracked version of cpplog to not break on any updates
as well as updates the CI build to the newer docker image versions.