Greatly improves clang precompilation time
Most of the compilation time used by clang (for "small" kernels) is consumed
to parse the clang default opencl-c.h header file.
Since this file does not change and we need it for all compilations, we now
precompile that header into a PCH separately and reuse the PCH for any successive
compilation of OpenCL C kernels.

Effects:
* Increases precompilation time of first/single kernels by up to 200ms
* Reduces precompilation time for emulation test cases from ~16.5s to ~4.5s!