Intrinsifies popcount()
This is mostly to support the llvm.ctpop() intrinsic, but since we already
do this, we intrinsify the whole popcount() calculation.
The generated code is up to 40 instructions smaller than the "native"
OpenCL C code, depending on the type.

Also fixes some read-after free and fixes out-of-bounds calculation for
memory area in emulator.