Adds built-in SPIR-V lexer
Adds a leightweight lexer on top of the existing SPIRVParserBase to
directly read SPIR-V binary inputs.

This allows us to support reading of SPIR-V binary inputs without the
need to compile and include the huge SPIR-V Tools libraries.

This is especially useful for providing the ability to compile SPIR-V
as intermediate language when used as compiler in the VC4CL library,
e.g. to support clCreateProgramWithIL and similar OpenCL C runtime
functions.