Moves precompiled modules out of VC4CL std-lib header directory
Now builds and uses the precompiled VC4CLStdLib modules from
<install prefix>/share/vc4cc. This should among others remove the warning
when deinstalling vc4cl-stdlib that the folder is not empty.

For same architecture (non cross-compilation) packages builds, the precompiled
modules are now also included in the Debian package. For safety (so they are
guaranteed to match the used LLVM version) they are still overwritten on
package installation.