Adds simple peephole optimization
There optimizations are run after any local is mapped to a register,
but before the labels are mapped to offsets and the intermediate
instructions are lowered to assembler instructions.

Thus, there optimizations have a "more complete" view on the generated
assembler, but in return MUST NOT modify any register mapping!

- Removes simple moves where input and output locals are mapped to the same register
- Removes simple moves where input and output locals are mapped to the same register file,
  if the readers of the output locals allow for it
- Removes NOPs inserted for physical register file read-after-write delay where the
  written local is mapped to an accumulator and therefore the NOP is not required