Adds comparison analysis
This analysis tries to find the information of the original comparison
responsible for writing a boolean value or setting a specific flags.

This will be later used to extend on the comparison detection for control
flow loop iteration information for induction variables.

Also extends pattern matching to allow gaps writing locals used later on
in the pattern as well as adds an optional flag to return the end of the match.