Reimplement fdiv with Goldschmidt's algorithm
Using a combinations of Goldschmidt's and Newton-Raphson
algorithms with argument reduction gives us an accurate implementation
with just 2 iteration steps.

Additionally, (u)short and (u)char division and remainder functions
are now implemented as derivated of the fdiv intrinsic,
which is accurate enough for numbers in the range 2^-16 to 2^16.

Effects:
* Reduces instruction count for short division emulation test from
  570 to 115. Similar for all (u)short and (u)char divisions and modulos.
* Fixes OpenCL-CTS/math_brute_force/divide -w (Wimpy mode)