Rewrite frontend source handling
Replaces file/stream-pair with abstract CompilationData structure to allow
for transparent handling of precompilation sources and outputs.
This will e.g. reduce the number of temorary files and reads/writes required
as well as allow for CompilationData structures which contain custom representations
of the data (e.g. LLVM modules stored in-memory as such) to be used later on.

This change breaks the API, but not functionality that is used by the VC4CL
runtime library.

Also replaces some #macro usages with C++ constructs to produce cleaner code
and allow for partial compiler syntax checking of disabled code branches.

Additionally unifies some LLVM code which is used by the BitcodeReader as well
as future clang/LLVM library precompilation steps and removes old clang library
precompilation which actually never worked.